Rename App to LegacyApp
