+++ site/src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './LegacyApp';
 import './style.css';
 import 'bootstrap/dist/css/bootstrap.css';
